Filter Journal by activity IDs of a given container

+++ src/modules/Elsa.Studio.Workflows/Pages/WorkflowDefinitions/Edit/ActivityProperties/Tabs/VersionTab.razor.cs
@@ -60,7 +60,7 @@ public partial class VersionTab
         return new TableData<WorkflowDefinitionSummary>
         {
             Items = response.Items,
-            TotalItems = response.TotalCount
+            TotalItems = (int)response.TotalCount
         };
     }
